Add mount trigger to inject script.
